BackEnd/feature_extraction_service/requirements.txt
fastapi
uvicorn
requests
bs4
lxml
selenium
dnspython